update: make Carousel resuable
