+++ src/components/carousel.module.css
@@ -1,0 +1,4 @@
+.carouselItem div, img {
+  width: 100%;
+  height: 100%;
+}
